working with Gordana to export maps as shape files as part of the loop in the code e.g. https://github.com/gomezcatalina/FishInverAtlas_Ricard/blob/master/FunctionsR/figure21.R
lines 107-122:
#### save shape files for FGP
species=as.character(dat.in$spec[1])
lname=paste0("SS",species,"_",yrs.labels[i],"_IDWmap-abundance")
path.FGP <- file.path(path.ATLAS, "FGP")
R_df <- as(R, "SpatialPolygonsDataFrame")
proj4string(R_df) <- CRS("+proj=longlat +ellps=WGS84 +no_defs")
 names(R_df@data)="legend"
 if(my.legend=="abundant"){   R_df@data$legend=c("0","<5","<20","<50","<100",">=100")

 }
 if(my.legend=="rare"){
   R_df@data$legend=c("0","<0.1","<0.5","<1","<5",">=5")
 }

writeOGR(R_df, dsn=file.path(path.FGP, "IDWMaps"),
layer=lname, driver="ESRI Shapefile", overwrite_layer=TRUE)
#### finish saving shape files for FGP

We tested this approach for Cod using Figure10.R & Figure11.R and it worked well.

Now testing for Lobster using Figure20.R & Figure21.R

The pdf outputs work fine, I can recreate the plots no problem for the new time periods. The issue arrives when I include the chunk to export the shape files. I am always able to explore the shape files as long as I do not assign values for the attribute tables. This is because I am defining 6 values for 6 rows of the attribute tables (e.g."0","<5","<20","<50","<100",">=100") and the attribute tables in my test (Lobster) has different rows depending of the time period. For example, using Figure21.R to plot IDW abundance I get the following rows in the attribute tables:

1999-2004: 6 rows
2005-2009: 6 rows
2010-2014: 5 rows
2015-2019: 5 rows

The legend assigned in the pdf is a text, independent of the number of polygons. So perhaps it is possible to have fewer polygons (for example if there is no "<5 " polygons?). But how could we assign the correct legend in these cases that number of rows in attribute tables may vary?

The number of rows does not change for Cod - I wonder if there are issues with the interpolation for lobster and in general for species that are not as frequent in some time periods, data is skewed in those cases, which may lead to skewed classification of the surface layers.